Rename digitalrf.py to digital_rf.py
